requirements.txt
kafka-python
psycopg2-binary
pandas
sqlalchemy
python-dotenv
pyyaml
